requirements.txt
torch>=1.12.0
torchvision>=0.15.2
Pillow>=9.4.0
pandas>=2.1.1
